# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/university/members/{membershipId}/handover": {
      "post": {
        "operationId": "UniversityStaffController_handover",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverUniversityMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffHandoverResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "University staff"
        ]
      }
    },
    "/api/v1/university/members/{membershipId}/disable": {
      "post": {
        "operationId": "UniversityStaffController_disable",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "University staff"
        ]
      }
    },
    "/api/v1/university/class-transfers": {
      "get": {
        "operationId": "UniversityStaffController_listPendingClassTransfers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassTransferRequestDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "University staff"
        ]
      }
    },
    "/api/v1/university/class-transfers/{id}/resolve": {
      "post": {
        "operationId": "UniversityStaffController_resolveClassTransfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveClassTransferDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassTransferRequestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "University staff"
        ]
      }
    },
    "/api/v1/account/me/class-transfers": {
      "get": {
        "operationId": "StudentClassTransferController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassTransferRequestDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Class transfer"
        ]
      },
      "post": {
        "operationId": "StudentClassTransferController_create",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassTransferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassTransferRequestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Class transfer"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zhixing Hub API",
    "description": "知行工坊模块化单体业务接口",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zhixing_session"
      }
    },
    "schemas": {
...
      "HandoverUniversityMemberDto": {
        "type": "object",
        "properties": {
          "successorMembershipId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "successorMembershipId"
        ]
      },
      "StaffHandoverResultDto": {
        "type": "object",
        "properties": {
          "fromMembershipId": {
            "type": "string",
            "format": "uuid"
          },
          "toMembershipId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "UNIVERSITY_DASHBOARD",
              "COLLEGE_ADMIN",
              "PROGRAM_LEAD",
              "COUNSELOR"
            ]
          },
          "reassignedClassCount": {
            "type": "number"
          }
        },
        "required": [
          "fromMembershipId",
          "toMembershipId",
          "role",
          "reassignedClassCount"
        ]
      },
      "ClassTransferRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "studentAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "fromClassId": {
            "type": "string",
            "format": "uuid"
          },
          "toClassId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_OUTGOING",
              "PENDING_INCOMING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "outgoingResolvedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "incomingResolvedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "resolvedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "studentAccountId",
          "fromClassId",
          "toClassId",
          "status",
          "createdAt",
          "outgoingResolvedAt",
          "incomingResolvedAt",
          "resolvedAt"
        ]
      },
      "ResolveClassTransferDto": {
        "type": "object",
        "properties": {
          "approve": {
            "type": "boolean"
          }
        },
        "required": [
          "approve"
        ]
      },
      "CreateClassTransferDto": {
        "type": "object",
        "properties": {
          "targetClassId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "targetClassId"
        ]
      },
...
    }
  }
}